Move package_setup_files to general lib to minimize code duplication
